add test_matrix and rename gplugins.eme/fem to gplugins.meow/femwell
