+++ .pre-commit-config.yaml
@@ -54,92 +54,3 @@ repos:
     hooks:
       - id: j2lint
         args: [--ignore, jinja-statements-delimiter, jinja-statements-indentation, --]
-  # - repo: https://github.com/pre-commit/mirrors-mypy
-  #   rev: "v1.0.1"
-  #   hooks:
-  #     - id: mypy
-  #       exclude: ^(docs/|example-plugin/|tests/fixtures)
-  #       additional_dependencies:
-  #         - "pydantic"
-  # - repo: https://github.com/terrencepreilly/darglint
-  #   rev: master
-  #   hooks:
-  #     - id: darglint
-  # - repo: https://github.com/pycqa/pydocstyle
-  #   rev: "a6fe4222012e990042c86cdaaa904a8d059752ee"
-  #   hooks:
-  #     - id: pydocstyle
-  #       additional_dependencies: ["pydocstyle[toml]"]
-  # - repo: https://github.com/asottile/reorder_python_imports
-  #   rev: 2b2f0c74acdb3de316e23ceb7dd0d7945c354050
-  #   hooks:
-  #     - id: reorder-python-imports
-  # - repo: https://github.com/PyCQA/pylint
-  #   rev: v2.14.1
-  #   hooks:
-  #     - id: pylint
-  #       args: [--exit-zero]
-  # - repo: https://github.com/macisamuele/language-formatters-pre-commit-hooks
-  #   rev: 6565d773ca281682d7062d4c0be74538cc474cc9
-  #   hooks:
-  #     - id: pretty-format-java
-  #       args: [--autofix]
-  #     - id: pretty-format-kotlin
-  #       args: [--autofix]
-  #     - id: pretty-format-yaml
-  #       args: [--autofix, --indent, "2"]
-  # - repo: https://github.com/adrienverge/yamllint.git
-  #   rev: v1.21.0 # or higher tag
-  #   hooks:
-  #       - id: yamllint
-  #         args: [--format, parsable, --strict]
-  # - repo: https://github.com/jumanjihouse/pre-commit-hook-yamlfmt
-  #   rev: 0.1.0 # or specific tag
-  #   hooks:
-  #       - id: yamlfmt
-  # - repo: https://github.com/pre-commit/pygrep-hooks
-  #   rev: "f6fb13e9516d1a9720a30a4049c930235abd642e"
-  #   hooks:
-  #     - id: python-check-blanket-noqa
-  #     - id: python-check-blanket-type-ignore
-  #     - id: python-no-log-warn
-  #     - id: python-no-eval
-  #     - id: python-use-type-annotations
-  #     - id: rst-backticks
-  #     - id: rst-directive-colons
-  #     - id: rst-inline-touching-normal
-  # - repo: https://github.com/nbQA-dev/nbQA
-  #   rev: 952c895a21aced4601165c43f77188559a5e825c
-  #   hooks:
-  #     - id: nbqa-black
-  #       additional_dependencies: [jupytext, black] # optional, only if you're using Jupytext
-  #     - id: nbqa-pyupgrade
-  #       args: ["--py37-plus"]
-      # - id: nbqa-flake8
-      # - id: nbqa-isort
-      #   args: ["--float-to-top"]
-  # - repo: https://github.com/hakancelik96/unimport
-  #   rev: 0.14.1
-  #   hooks:
-  #     - id: unimport
-  #       args: [--remove, --include-star-import]
-  # - repo: https://github.com/pycqa/isort
-  #   rev: "3a72e069635a865a92b8a0273aa829f630cbcd6f"
-  #   hooks:
-  #     - id: isort
-  #       files: "demo_project/.*"
-  #       args: ["--profile", "black", "--filter-files"]
-
-  # - repo: https://github.com/kynan/nbstripout
-  #   rev: e4c5b4dcbab4afa0b88b7a9243db42b1a8d95dde
-  #   hooks:
-  #     - id: nbstripout
-  #       files: ".ipynb"
-  # - repo: https://github.com/codespell-project/codespell
-  #   rev: 68ea3cff5eba266e0814794cce6cd19537f58137
-  #   hooks:
-  #     - id: codespell
-  # - repo: https://github.com/pre-commit/pygrep-hooks
-  #   rev: 7b4409161486c6956bb3206ce96db5d56731b1b9 # Use the ref you want to point at
-  #   hooks:
-  #     - id: python-use-type-annotations
