+++ gplugins/femwell/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .solve_thermal import solve_thermal
+
+__all__ = ("solve_thermal",)
