+++ gplugins/meow/__init__.py
@@ -1,0 +1,5 @@
+from gplugins.meow.meow_eme import MEOW
+
+__all__ = [
+    "MEOW",
+]
